Detect Composer Autoloader
Detect if Composer's autoload file can be find in the vendor directory in this project (if this project is the root package), otherwise assume that this project has been installed as a dependency and load the root package autoloader.
Also, change "include" to "require_once"; as the autoloader is required (once only) to load the ProcessSlave class.